+++ database/install/myems_system_db.sql
@@ -1274,6 +1274,6 @@ USE `myems_system_db`;
 INSERT INTO `myems_system_db`.`tbl_versions`
 (`id`, `version`, `release_date`)
 VALUES
-(1, '1.9.5', '2022-07-10');
+(1, '1.9.5', '2022-07-18');
 
 COMMIT;