+++ CHANGELOG.md
@@ -10,6 +10,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - added tbl_template_files to myems_reporting_db
 - added tbl_integrators to myems_system_db
 - added alarm audio for notification in myems-web
+- added alarm lottie for notification in myems-web
+
 ### Changed
 -