+++ CHANGELOG.md
@@ -11,7 +11,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - added tbl_integrators to myems_system_db
 - added alarm audio for notification in myems-web
 - added alarm lottie for notification in myems-web
-
+- added tbl_data_repair_files to myems_historical_db
+- 
 ### Changed
 -