+++ CHANGELOG.md
@@ -62,6 +62,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - Updated MeterTracking Excel exporter.
 - Updated MeterBatch Excel exporter.
 - Deleted width limit from cascader-menu in myems-web.
+- Updated falcon web framework version to the latest (3.1.0) in myems-api 
 
 ### Fixed
 - updated meterrealtime report to display right units of energy value.