+++ CHANGELOG.md
@@ -15,7 +15,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - added data repair file actions to myems-api
 - 
 ### Changed
-- 
+- changed parameter line chart to multiple line chart for meterenergy report in myems-web
 
 ### Fixed
 - removed invalid ORDER BY clause from myems-api