+++ CHANGELOG.md
@@ -5,6 +5,16 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## [Unreleased]
 ### Added
+- 
+### Changed
+-
+### Fixed
+-
+### Removed
+-
+
+## [v1.9.5] - 2022-07-18
+### Added
 - added subtotal to meter/offlinemeter/virtualmeter batch report 
 - added mqtt to data source protocol in myems-api
 - added tbl_template_files to myems_reporting_db
@@ -896,7 +906,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ### Removed
 -   None.
 
-[Unreleased]: https://github.com/MyEMS/myems/compare/v1.9.4...HEAD
+[Unreleased]: https://github.com/MyEMS/myems/compare/v1.9.5...HEAD
+[v1.9.5]: https://github.com/MyEMS/myems/compare/v1.9.4...v1.9.5
 [v1.9.4]: https://github.com/MyEMS/myems/compare/v1.9.3...v1.9.4
 [v1.9.3]: https://github.com/MyEMS/myems/compare/v1.9.2...v1.9.3
 [v1.9.2]: https://github.com/MyEMS/myems/compare/v1.9.1...v1.9.2